requirement.txt
matplot==0.1.9
numpy==1.18.5
pandas==1.0.5
Keras-Preprocessing==1.1.2
seaborn==0.10.1
